[improvement] Add BlacklistingChannel (#97)
<!-- PR title should start with '[fix]', '[improvement]' or '[break]' if this PR would cause a patch, minor or major SemVer bump. Omit the prefix if this PR doesn't warrant a standalone release. -->

## Before this PR
Errors seen in a channel would not affect future decisions to use that channel.

## After this PR
After an error, we will wait for some cooldown period before using that Channel again.

## Other Considerations
We don't currently have any mechanism for forcing requests to execute when none of the underlying `LimitedChannels` have capacity. That seems desirable, though we may want some escape hatch in the BlacklistingChannel to reset them when all Channels have been blacklisted.