Fix dialogue channel tracing
Previously if no active trace existed, outgoing requets did not
contain trace IDs.
Normalized all tests to use Channels.create to validate all
client implementations handle tracing correctly. Note that
the jdk9 client was tested both with and without the Channels
wrapper, now it is only tested with.
We never expect channels to be used without the wrapper, so it's
more important to test the wrapped functionality than unwrapped.