fix #406: QueuedChannel does not leak responses in edge cases
Updated to use the more scalable ConcurrentLinkedQueue with a
separate variable to track the estimated size. This allows us to
avoid decrementing after attempting to send a request until we
know the request has not been rejected. This also allows us
to make a lock-free read when events are enqueued to avoid the
fast path when the queue is not empty.