Dialogue avoids wrapping all runtime exceptions for cleaner results
In converting to dialogue I found several tests failed, even using
blocking clients, because the expected exception was wrapped in
an UncheckedExecutionException.
We can allow runtime exceptions but attach a suppressed exception
with the stack trace of the rewrap to avoid the complexity of
supporting re-creating all types while preserving the original
information.