POC for sticky sessioning
Will discuss internally, but we have a lot of garbage at present related
to handling load balancing - we need a 'session' of requests to an
underlying channel.

Wondered if this might be cleaner?

Adds a 'tag' concept to requests and responses, where channels can add
and remove data as relevant (although the present implementation of
'getTag()'.

Then, uses the tag concept to implement a limited channel which can be
used to implement sticky sessioning.