PinUntilError stable hostindex metrics (#689)
The `dialogue.pinuntilerror.success` metric had a `hostIndex` tag which was pretty useless because it was the index into the _shuffled_ list of nodes.  Now, this index is preserved from the _original_ list of nodes, so we should expect to see a square wave style pattern as nodes are reshuffled every 10 minutes.