Fix a thread leak from unclosed httpclients
The CloseableClient object closes delegate clients
when it is garbage collected, instead of leaking.

The wrapper object is only released once it's no longer used
because it's referenced via the channelFactory object and
the cache, so as long as a service is making requests the
dialogue channel is pinned, and keeps the client around.

Longer term we may wish to implement reference counting
to avoid finalizers and release resources as soon as the last
request completes.