Idle connections are checked and timed out more frequently
Previously the idle timeout and check interval were the same,
resulting in the potential for connections to live twice as long
as we expected. While this isn't a big problem, it introduces a
potential race when the server closes an idle connection as the
client attempts to check it out from the pool. The liveness check
can succeed while the subsequent request occurs after the
server-side has closed.