Dialogue apache clients use the read timeout as socket timeout
Previously we used the maximum value which broke consumers with
reduced read timeout in order to prevent excessive blocking.

Apache HttpClient does not support separate read and write
timeouts.

Risks: It's possible that some services rely on a modified write
timeout which would have similar problems, but this scenario is
much less common.

Alternatives considered: I thought about using the minimum of
read and write timeout, but that runs into similar problems when
the read timeout is increased.