Handshake time is no longer bounded by the connect timeout
Large garbage collections cause handshakes to fail. Ideally we
would not take more than a few hundred milliseconds to handshake,
but that's clearly not the current state of the world.

Apache ConnectTimeoutException is mapped to a more specific type

Currently our retry metrics show `safeioexception` which isn't
terribly informative. This change will replace those with the
self-describing `safeconnecttimeoutexception`.