Report the differebnce between client and server request duration
This approach leverages the connection `receiveResponseHeader`
method which is as close to the socket as we can get, but has
the drawback that it cannot account for headers sent earlier
with large request entities.
In many cases the request header and entity can fit into a single
packet, so this is a reasonable tradeoff, and captures data for
our most performance sensitive endpoints.