Handshakes time out based on the connect timeout
Note that the connect timeout isn't always used directly, only
if it's ten seconds or greater. Otherwise, the minimum of ten seconds
and the configured socket timeout is used. This allows us to
retry on another host when handshakes are blocked (gc storm on
the target) without risking overwhelming hosts when the connect
timeout is set to an unusually low value.