Dialogue allows the default port to be used
This behavior is based on the configured URI -- URIs which provide
a port result in a `Host` header containing a port, and those which
rely on the implicit default (e.g. http -> 80, https -> 443) do
not include the port in their `Host` header.

It turns out some servers require that the default port is used,
and _not_ included in URIs. Until we have a better way to test this
I've opted not to exclude default ports everywhere, only when
the provided URI doesn't include a port. Testing is difficult because
permissions are required to run a server on port 80 or 443 to verify,
however I have validated that this behaves in the way we expect
locally.