fix #1308: Reduce Dialogue client creation overhead
This bounds work to the number of endpoints that are actually
used. In most cases only a small subset of endpoints are
exercised, this is particularly true in many sticky-session
use-cases in which instances are created much more frequently.

Note that our serializers and deserializers can be relatively
expensive, however the expensive components are already
memoized. Per-endpoint channel binding attempts to do additional
work up-front, however it means each client instance we create
was doing _all_ of this work immediately. Now it occurs lazily
the first time an endpoint is used for a more predictable
performance profile.