Delete the undocumented `dialogueVersion` metric tag
This tag was added prior to library-version tags for each
library which used metric-schema. Because `dialogueVersion` was
not included in metric-shema, it was not documented for consumers
to rely upon.
Adding this tag to all metrics resulted in non-trivial overhead,
especially for sticky-channels where channels are re-created in
hot paths.