Java 11 clients use TLSv1.3 in addition to TLSv1.2
Previously TLSv1.3 was gated on JDK-15+ due to issues in early
releases of java 11. These have been resolved in later releases,
at this point we want to allow all clients to use modern protocols
to give us the option to remove TLSv1.2 support as early as possible
in case vulnerabilities are discovered.