Lazily initialize empty instance deserializers
Encodings are alreay lazy to avoid up-front work across endpoints
that may not be used, however the empty instance factory will
attempt to use jackson to deserialize a null node into the requested
type immediately. This isn't supported by all endpoints, and is
as expensive as the first time an endpoint is requested. Attempting
to do this for all return values for all endpoints doesn't make
much sense.