The first retry does not incur a backoff penalty
This could be predicated in some way on the total number of nodes,
however multiple successive failures should be unlikely.
In most cases, a retry onto a healthy node as quickly as possible
is ideal. There may be cases in which this increases burst load,
however the risk/reward tradeoff for user-facing latency seems
reasonable here.