RetryOtherValidatingChannel handles mesh URIs
In a previous change, we removed special handling of mesh uris
because the configuration URIs weren't used in practice, in favor
of the TargetUris list, which removes mesh-prefixes automatically.
One edge case we failed to deal with was mesh-uris impact on
RetryOtherValidatingChannel. We've updated the implementation to handle
mesh-uris, and out of an abundance of caution opted to strip mesh
prefixes from the configuration used elsewhere to avoid potential impact.