fix #1076: Dialogue allows colons in http paths and query parameters
This is required by some GCP APIs, where encoded colons are not
respected.
[rfc3986 section 3.3](https://datatracker.ietf.org/doc/html/rfc3986#section-3.3)
clearly allows colon as a pchar, however not all servers are necessarily fully
compliant with the rfc. We've had issues with specific sub-delimiters
in the past. This shouldn't be terribly risky, esepcially given this
implementaiton has never allowed unencoded colons before, and we have
no evidence to suggest they should be problematic.