Implement support for the Proxy-Upstream-Request-Attempts header
This aims to provide coordination between clients and reverse
proxy servers which implement retries, allowing us to avoid
multiplying out the total retry count.

When dialogue is configured with more retries than a proxy, we still
may retry more than we want, but not DIALOGUE_ATTEMPTS * PROXY_ATTEMPTS.
The worst case scenario should be when the proxy retry attempts don't quite
reach the configured dialogue retry limit, and another full round of proxy
retries must be accumulated. This should be a reaosnable trade-off because
it allows the calling client to be configured with a larger value as
needed for stability, without multiplying all retries across the board
unnecessarily.

Note that when a Dialogue client receives a tcp error/closure, it
can make no assumptions about proxy retry attempts, so this only
applies to requests which successfully receive a response.