Dialogue allows unencoded equals signs (=) in URL query parameter values (#2574)
Dialogue allows unencoded equals signs (=) in URL query parameter values

This is not an API break, however we're using a breaking changelog entry for visibility in case of unknown non-compliant servers.

Previously the `=` character would be encoded as `%3D`, which is also allowed, but not required, by rfc3986. Some server implementations require equals signs in query parameters not to be encoded.

This encoding is an implementation detail within dialogue, where either way is valid for servers which are compliant with the rfc.

It is possible that some servers or proxies do not handle unencoded equals signs correctly. Please reach out to us if you find cases where servers do not behave as expected.