alternative implementation of deadline enforcement
ReloadingClientFactory.ReloadingParams is modified to support an
Optional<Boolean> indicating deadline enforcement. When a value is
present, this dictates how we construct a DialogueChannelFactory which
is passed to DialogueChannel's builder and used to construct channels.

This change adds a utility class DeadlineEnforcement which has a method
for constructing a DialogueChannelFactory that wraps an existing factory
and, dependeing on the value of the Optional<Boolean> from the reloading
params, creates a channel which sets a request attachment indicating the
enforcement strategy; this then wraps the channel created by the
delegate factory.

DeadlineAdvertisementChannel looks for the request attachment to
ultimately decide the deadline enforcement strategy to use.

Ideally the DeadlineEnforcement class could be made package-private, but
dialogue-clients needs visibility to create the delegating
DialogueChannelFactory, and DeadlineAdvertisementChannel also needs
visibility in order to read the request attachment.